migrate import: ensure all files end up in .gitattributes
Currently, if we have a file and it contains an extension, we'll add
that extension to the list of paths that should be placed in
.gitattributes.  However, if the file does not contain an extension, it
won't be listed at all.  As a result, we may convert a file flagged by
the --above option but never cause Git to filter it properly.

Let's fix this by writing an attribute into the repository with the path
relative to the root if we don't contain an extension.  This means that
people can store typical Unix binaries, which lack extensions, as LFS
files more easily.


Former-commit-id: dd463c624fbe90ce0492a5b8a19e85a1f3843d8b