git: use pktline repository
We have a pktline repository that provides our pkt-line code from Git
LFS, but separated out into a separate, reusable repository.  The
current code is private to the git package, and since we'll want to
use it in the future from multiple packages, it makes sense to just use
the separate repository, which we want to do anyway.

Let's include the repository in our go.mod and vendor the modules.
Remove the old code and replace its use with the new code.


Former-commit-id: affd5e9f1c9a22240ec4849ea3da3140f2da9bae