docs: add a quit command to the SSH protocol
On Unix systems, it's easy for a server spawned by an OpenSSH instance
to determine when the other side has gone away: it gets an EOF or a
SIGPIPE.  However, on Windows, this isn't the case, since SIGPIPE
doesn't exist, so it's not necessarily easy to determine if the remote
side has hung up.

To help everyone involved keep things simple and straightforward,
regardless of operating system, and to distinguish expected termination
from unexpected termination, let's add a quit command.  This is the
expected and normal way to shut down a connection.


Former-commit-id: 18c15dea193ad710b86a26ad9c256177837b1e67