docs: remove "lock" operation from SSH protocol
Right now, the "lock" operation on the command line can be used to
perform both read-only and read-write operations.  This makes
authentication tricky, since it's not easy to determine from
command-line arguments which permissions should be used.

To make things simpler, let's use the commands "upload" and "download"
like they are for git-lfs-authenticate to adjust locking.


Former-commit-id: c3bed437a385bd450e70dd321a90dd3038732142