lfs: don't invoke diff drivers when pruning repositories
Certain file paths are associated with diff textconv patterns with Git
for Windows, and these are invoked by default when using git log.
However, when we're using git log to scan data to find LFS objects, we
don't want to invoke those programs.

Let's fix this by passing appropriate options to disabled these
invocations.  In the test, we set the driver to false, since Git will
fail if the process it calls also fails.


Former-commit-id: fcb163272bbb16d108b7d6e162e3485e2f4e5330