.github/workflows: switch to Ubuntu 20.04
Ubuntu 20.04 is the latest LTS version and it provides the newest
software available for us to build.  More importantly, it appears that
the latest Go, 1.16, will be available for Ubuntu 20.04 but possibly not
for Ubuntu 18.04, so because we want at least that version in order to
successfully build for ARM64 macOS, let's prefer the latest Ubuntu.


Former-commit-id: 282ac68cc6ad373279a986b22f285147e7a180da