.github/workflows: use Go 1.16 anywhere we "make release"
When we run "make release", we will implicitly build all possible
release targets, including ARM64 Darwin.  However, ARM64 Darwin is
available only in Go 1.16, so if we try to build with an earlier
version, the operation will fail.  For now, let's be explicit about the
version we want to use so that we don't have to rely on the default
Go version on Actions.  This will let us get ARM64 Darwin support out
faster.

In the future, when the default for Actions is Go 1.16 or later, we'll
probably revert this commit and go back to using the default behavior.


Former-commit-id: 902ee708213ab0d61fb6629cc7b050b8f96da788