Use tools.CanonicalizeSystemPath to canonicalize paths
Now that we have a function which does path canonicalization correctly
on a variety of systems, let's use it in favor of filepath.EvalSymlinks.
We use this function to handle any path we know doesn't come from
outside Git LFS, since we know we'll never need to handle Cygwin paths
in these cases and it's more efficient not to invoke a subprocess if
it's not necessary.


Former-commit-id: b6f0613f073a811507fe2522e62349c9904e9388