commands,t: add migrate info --top option tests
We can simplify one line in the migrateInfoCommand()
function where it resets the "entries" list of informational
entries to be output to include no more entries than the
setting of the --top option, because we have already
clamped the migrateInfoTopN value to be between zero and
the total number of entries.  So we can drop the second
call to tools.MaxInt() here; there is no chance
migrateInfoTopN could be less than zero at this point.

To confirm our changes and any future ones work as
advertised, we then add one test and revise one other.

The new test checks that the --top option works in
multiple cases, including when its value is zero or below
zero, or when it is greater than the number of available
entries to be printed by the "git lfs migrate info" command.

The revised test is changed so that it confirms that only
the highest available entry is printed when --top=1 is
specified.  Previously, because the --above option is also
set in this test, there was only one matching entry to be
printed, and so the --top=1 was redundant.  If --top=2 had
been specified, the test would still have passed.


Former-commit-id: b8f8686c6b959176aa4442d1fdb7f370e68e8c46