tq: implement a pure SSH-based protocol for transfers
A pure SSH-based protocol has been a long time request from many users,
so let's implement one.  Implement basic upload and download support,
plus batch requests, and support pkt-line tracing for text packets to
make debugging easier.

Note that locking is not yet supported; that will come in a future
patch.

We prefer the Endpoint function to the RemoteEndpoint function because
the former handles lfs.url and the latter does not.

Update a comment about the shared temporary directory; it is no longer
cleared automatically by the adapter, and instead the cleanup happens
later in other code.  Therefore, it is safe to share the directory among
the transport adapters.

Co-authored-by: Chris Darroch <chrisd8088@github.com>

Former-commit-id: 594f8e386cce3441e06c9094ab5e251f0e07ca1f