t: test additional stashed objects with prune
In each of our tests of the "git lfs prune" command
when it follows the "git stash" command, we add LFS
objects on a branch and stash them before the branch
is deleted.  In all cases we expect the additional
stashed objects to be retained by "git lfs prune"
while unreferenced objects on the deleted branch are
pruned.

Note that these chnanges also ensure we are test
the "git lfs prune" command with multiple stashed
entries rather than just a single stash entry.


Former-commit-id: fbac6e0a905680466e1cade9d8c548a9aed4ce6a