config: skip cleanup if config struct is nil
If a SIGINT is delivered before the commands.cfg Configuration
structure is initialized then when the goroutine which responds
to the signal runs the commands.Cleanup() method, it panics after
dereferencing a null pointer to the Configuration structure while
trying to acquire the mutex that protects the struct's fs member.

We can resolve this by simply testing for a nil Configuration
structure pointer in the config.Cleanup() method and returning
immediately in that case.


Former-commit-id: c7d2a6b89eb9b0f33bf8a140e043b629e1b6364d