docs: make owner-id optional for locks in SSH protocol
Ideally, we'd be able to wrap a server speaking the HTTP protocol with
the SSH client for easy backwards compatibility.  However, our current
list-lock command, which is used for the lock listing and verifying
endpoints, always emits the ours/theirs indicator, which is not
available in the HTTP protocol unless the user has push access.

To make this kind of bridging possible, let's specifically mention that
the owner-id data is optional in the download interface, but is required
in the upload interface.


Former-commit-id: a9edd9490553893de7279b09f8866a5d9b429779