commands,lfs: add migrate info --pointers option
At present, the "git lfs migrate info" command reports
the total sizes, numbers, and file extensions for blobs
which have already been converted to LFS pointers, and
it uses the size of the pointer blobs themselves instead
of the size of object data referenced by the pointers.

The result is a confusing output which does not distinguish
these types, has misleading size totals, and generates
output listings even for repositories fully migrated to LFS.

We therefore add a new --pointers option to the "migrate info"
command so that a summary of all valid LFS pointers found in
the repository traversal may be either skipped entirely or
reported separately.

To support this option we add an lfs.DecodePointerFromBlob()
function which parallels the existing lfs.DecodePointerFromFile()
function.

We also refactor some of the entry lookup logic in the
BlobFn callback function used in the migrateInfoCommand()
function into a new findEntryByExtension() utility function.


Former-commit-id: 9c774cef385a209c88603db226ed90884a9fd658