t/t-migrate-info.sh: split out --pointers tests
We split out the existing tests of various values of the
"git lfs migrate info" --pointers option, added in PR #4436,
in order to treat each value in a separate test.  This has
the small advantage that if support for a particular option
value should be broken unexpectedly in the future, it
should result in a specific test failure instead of a more
generic one.


Former-commit-id: b386d39dee3b953a85b437d3ef4e8e326e9f8e88