migrate import: make --above affect only individual files
The --above option is supposed to only affect individual files, but in
its current configuration, the entry added to the .gitattributes file is
a wildcard pattern for the entire extension.  This is a problem because
other files with the same extension won't necessarily be converted,
leading to the dreaded message that some files "should have been
pointers, but weren't".

Let's fix this by doing what the documentation says: only adding
individual files themselves.  That fixes this problem and also the
inconsistency in our documentation.


Former-commit-id: 8490bf31483a06db435774358373053e3fcee5fd