commands/pointer: add an option for strict checking
Currently we process pointers in a rather lax way and accept a variety
of non-canonical pointers, such as those with CRLF line endings or a
size value of 0 (which should be an empty file instead).  However, we
have a flag to indicate whether the pointer is canonical when parsing
it, so let's add a --strict flag so that we can allow people to reject
non-canonical pointers.

Note that while strictness is not the default, we add a --no-strict flag
as well so we can allow people to use it now and change the default at a
later time (probably 4.0).


Former-commit-id: 552f7073af742714cccf8cd6557265b6466c2dd1