lfs: make lsTreeBlobs more generic
Right now, this function lets us check whether a file is a pointer or
not.  However, in the future, we'll also want to use this function to
find all of the .gitattributes files as well as pointers.  Let's make
the function take an arbitrary predicate so we can reuse it for both
purposes.


Former-commit-id: 4c0e51ff7079f4999b2c045569bab26b8ea2ece5