git: allow reading attributes from an arbitrary Reader
At this point, we only read attributes from files on disk.  However, in
the future, we'll need to handle attribute data being read from the
output of git cat-file --batch.

In order to do so, we'll need to handle reading from an arbitrary
Reader, so let's refactor out that code into a separate method and make
the file reading code handle it.


Former-commit-id: 6bfbde868ad254cf98d0d3177fb3bf93a5aef806