commands/fsck: add a check for pointer files
Now that we have support for checking invalid pointer files, let's add
an additional check for that in git lfs fsck.  Emit a message for each
pointer file which is non-canonical and each file which should have been
a pointer but was not.

Add options to control which checks are run, and default to the old
behavior for compatibility.  While we're updating our code for multiple
checks, let's also avoid relying on the fact that our check functions
return nil when there are no problems and instead rely on them returning
any empty slice, whether nil or non-nil.

Document these new options in the manual page as well.


Former-commit-id: 8c12e4802a66b533b29afc82d18f2378f789b624