commands/fsck: allow scanning revision ranges
Currently, git lfs fsck operates only on the HEAD and index.  For
objects, this is usually the right decision, since only objects for HEAD
are checked out.

However, for pointers, this may not be desired.  A user may want to
check pointers for a range of commits, such as during a CI job.  To deal
with these cases, let a user specify a revision or a simple range of
revisions to operate on, and process those revisions.

Note that we don't currently process the index with --pointers because
this requires a completely different set of scanners which are not yet
implemented.  We can implement such a feature in a future revision if
desired.

In the tests, refactor out our setup code into a function for reuse in
multiple tests.


Former-commit-id: e6c9d1de195dbc67eecb58505ab6898828e3db8a