locking: reject multiple paths
Currently, when trying to lock or unlock multiple paths, we simply
ignore paths other than the first.  This is surprising to users, so
let's print an error in this case.  In addition, let's have both
commands exit unsuccessfully to indicate failure.

Note that we don't just allow multiple paths because the JSON format
can't handle multiple items.


Former-commit-id: 9fad6ff488f4446a24bd9769b4b6fc4edde07673