lock: handle absolute paths properly
Currently, when someone passes an absolute path to git lfs lock, we end
up stripping off the leading slash, appending the directory name a
second time, and then using that as a relative path.  Obviously this is
nonsensical, so let's instead canonicalize the path if it is absolute
and then make it relative.

We canonicalize the path first because our repository path will always
be canonicalized, and doing so means that we will end up with a relative
path within the repository even if the working directory or the absolute
path given contain symlinks or other non-canonical forms.


Former-commit-id: 8842816d279c0ca7b9cef903a4a95fee1cacf181