tq: drop unused ClearTempStorage adapter method
The ClearTempStorage() method of the transfer Adapter interface,
while implemented for most adapters, is never actually called,
and moreover is dangerous in some its current implementations
if it were ever to be called.

Specifically in the basic upload and download adapters and the
SSH adapter, the ClearTempStorage() method relies on an internal
tempDir() method to return the path to the temporary directory
to be cleared, which ClearTempStorage() then removes entirely.

The tempDir() methods, however, can return os.TempDir() if they
are unable to create a local temporary directory within the
Git LFS storage directory, and also do not cache the path to
the directory they first returned.  So were ClearTempStorage()
ever to be called in one of these adapters, its invocation of
tempDir() might just return the system temporary directory
(e.g., /tmp on Unix), which it would then attempt to remove.

The ClearTempStorage() method was introduced in commits
8d7525d533f5b9ba8b00d8d9f512b959aa609c14 [formerly f124f0585f73279195cb20ecb4aa9e149512b48a] and
8689cd6bd4850b291241ec2e165ae8d18d75bb5a [formerly 940a91a8df18eaf16e9a0162999f26d59fa9f5af] of PR #1265, but
has never been called by any user of the transfer
adapter interface.

We therefore just remove this method and its implemetations
from all the current tranfer adapters.


Former-commit-id: e1624ca68524e53c15fce27f0dcfa5f4b066d95f