checkout: rewrite path specs on command line
When a user gives a path spec to git lfs checkout on the command line,
we turn it into a pattern and then check out only the files specified.
However, currently the user can specify a directory and things will
work, but in the future, we won't automatically rewrite paths for our
file path filter instances.

To handle this gracefully, let's rewrite paths that refer to a literal
directory as paths referring to every path underneath them.  This lets
users write "." and get everything in the current directory, which is
probably what they wanted.


Former-commit-id: 56abb7122a7aa62408d0a14e23907d88b431abdf