Add tests coverage for retry-after batch endpoint
This commit adds integration test coverage for the LFS batch API endpoint using
the 429 retry-after scenario for both the upload as well as download case. It does so for both single- as well as multi-file scenarios to validate the proper batch queueing of multiple failed OID requests. It furthermore moves the existing storage API endpoint
tests to for the same scenario to a more logical `t-batch-storage-retries-ratelimit.sh`
file to not cause confusion with the actual batch API endpoint tests.


Former-commit-id: 32f571d81d56c1f21977b67bcfd42a14f651c4e4