t: test fetch/prune diff parsing with ANSI escapes
We set the "color.diff" Git configuration directive to "always"
in several tests which exercise the scanning and parsing of
"git log -p" output in order to validate that they still
succeed even with ANSI color code escape sequences in the
Git diff output.

Unfortunately, at the present time, the tests will no longer
succeed as a result of this change, but we intend to resolve
that problem in a subsequent commit.


Former-commit-id: 2752acc925d565446ef954a6094cb05aa7ac6e2d