lfs/gitscanner_log.go: disable color for log diffs
We explicitly set the --color option to "never" when running
"git log -p" and parsing its output; this ensures that we do
not include ANSI escape sequences in the file names captured
by the fileHeaderRegex.  If we fail to do this and the user
has set the "color.diff" Git configuration option to "always"
then we will incorrectly parse the Git log output and report
"Unable to parse pointer from log" errors.


Former-commit-id: 1fbbf3d74da24371822c77535e26b18ae5226b88