lfs/gitscanner_log.go: anchor diff line regexps
We can somewhat improve the speed with which we parse the
output of "git log -p" commands by anchoring the two regular
expressions that match lines beginning with "diff" so that
they no longer need to scan the entire line.

This is only valid because in a previous commit we ensured
that no ANSI escape sequences can now appear in the diff
portions of the Git log output, and further, because we never
supply the --line-prefix option.  Without that option and
with diff colorization forced off, all "diff" headers should
always begin at the start of a new line.

(Note that the fileMergeHeaderRegex is not actually used as
we never supply the --cc option to "git log", but because
we may make this functional in a future change, we update this
regexp as well as the actively utilized fileHeaderRegex one.)


Former-commit-id: e51bb972f4afbf6f6fa6f2886a107c913e9a4098