Stop supporting Go older than 1.13
When using Go 1.12, we need to download some of the HTTP/2 code as part
of the build process.  However, that code uses the Go 1.13 errors
package and expects errors.Is to work, which it does not.  Since this
isn't really practically avoidable and Go upstream is not going to care
about getting those older Go versions to work, let's drop support for
Go before 1.13.

Note that we only guarantee compatibility with the latest Go version, so
this doesn't violate our support policy.


Former-commit-id: 7760b34d43b3eff317dfaedcbfcd14b39b2560c4