ls-files: don't process empty files as pointers
In 3.0, we started processing empty files as pointers, since they are
their own pointer files.  However, this caused git lfs ls-files to
process empty files that are not tracked by LFS as pointer files as
well, which breaks various things.

Ideally, we'd filter these out such that empty LFS files are pointers
and others are not, but this is not efficiently implementable with our
current codebase, so let's revert to the old approach.


Former-commit-id: 5a5b7c4816b2ebe53a69481db8333d818bce2a49