git: honor GIT_OBJECT_DIRECTORY
When we're in certain server-side hooks, the GIT_OBJECT_DIRECTORY
environment variable is set so we can read and write temporary objects
which are quarantined there.  However, Git LFS doesn't currently honor
that environment variable.  Let's make it do so, so that we can invoke
Git LFS in the normal way in such an environment.


Former-commit-id: a03d55713a210d72a1a4a283b584384f51bb1347