lfs: don't flag non-LFS files as invalid pointers
When scanning the repository for invalid pointers, we set up a filter
with all the paths from .gitattributes for that commit.  However, the
default behavior of a filter with no paths is to return true for all
objects.  As a result, if there is no .gitattributes file, or if it
contains no filter=lfs lines, then we flag all files in the repository
as LFS files and complain about all of them not being pointers.

Since this isn't the desired behavior, let's correct it by setting the
default behavior of the filter to false and add a test for this case.


Former-commit-id: 90131475835a7f0963b3e86bfc0602d070a43632