script/upload: don't require empty bin/releases for finalize
When the upload script is used in finalize mode, which is a manual part
of the release process, it currently requires the bin/releases directory
to be empty.  This is because it uses the upload_assets function, which
uploads everything in that directory, and we obviously don't want to
upload whatever the developer may have in that directory as part of our
release, only the signed hashes.

Since the person doing the release is usually the person who has built
the pre-release assets for the pull request, this means that we
invariably bother this person, which, while not the end of the world, is
a minor annoyance.  Let's make this experience better for the core team
member doing the release and lift this restriction.

Make the upload_assets function take a location of assets to upload, and
when we're uploading assets in non-finalize mode (which is now
effectively restricted to our release workflow) use the bin/releases
directory, but use a temporary directory for the finalize step.  Remove
the check an empty bin/releases directory, since we no longer need it.


Former-commit-id: e40219fcb05c499f2369a44370c3af9a0ffb940f