git: keep reading config if not in a Git repository
If we're not in a Git repository, the call to IsBare will fail with an
error.  In such a case, we promptly return and fail to return any of the
configuration we did read.  Let's fix this by just skipping the read of
the .lfsconfig file and returning the configuration we're already read.


Former-commit-id: a10748f18ee44e9a54ca1d11675f754fa94ce5e3