errors: add a function to extract exit status from error
Right now, we have some custom code to extract the exit status from an
appropriate error.  However, we'll need to use this in multiple places,
so let's refactor it out into a function.

We use the Go 1.13 errors package because we no longer test for Go 1.13
since our dependencies also use this package and there's no practical
way to avoid this dependency.


Former-commit-id: 44b8b6aba6a5cb8bdfa5c15e1d5bfc3b5676193a