.github/workflows: include gettext when building
When building the default image in CI, or a release build, ensure that
the gettext package is installed so we can test locales.  We don't
include it in all CI runs to ensure that we also test building without
localization.

Note that on Windows, these tools should always be installed as part of
the Git for Windows environment, since they are needed for Git, and we
would have no way to install them anyway.


Former-commit-id: 628571695fde7ec92b6d18ae87d8d7e4dc87d5d4