docs/man: remove parentheses from headers
As noted in a prior commit, the ronn(1) utility does not
fully support the use of parentheses in headers, and as a
result fails to properly format references to such headers in
its roff output and generates broken anchors and references
to them in its HTML output.

We therefore just rewrite our manual pages' headers and all
references to them so they do not contain any unsupported
characters like parentheses or slashes.  This also has the
advantage of making some of these headers more explicit and
descriptive.


Former-commit-id: 412ca7d4f400df9a9d91a7f231c7db593a975f2b