t/t-checkout.sh: expand conflicts test
In a previous commit we added an error message to the
"git lfs checkout" command which will be output in certain
cases when the --to option is supplied and either zero or
multiple paths are provided to be checked out.

To validate the previous commit's change we expand our
"checkout: conficts" test to check the relevant cases,
and we also add some checks for other invalid options
and arguments, such as when no "stage" option is given
(--base, --ours, or --theirs), when too many "stage" options
are set, or when paths to conflicting files are provided
which do not correspond to Git LFS objects.

Note that while our new checks will not catch a Go panic
in particular, they would still cause the test to fail
if our previous commit's change was not in place, because
the expected new error message would not appear.


Former-commit-id: 7da65df58630641ae279a760f7ceeb21ffef93a3