filter-process: don't print large file warning on fixed versions
In Git for Windows 2.34.0 and the upcoming Git 2.35.0, the issue that's
caused corruption on Windows has been fixed.  Let's avoid warning users
that there's a problem if they're using 2.34.0 or newer unless they've
specifically set the option to true.  Note that we only consider 2.34.0
and not 2.35 since Unix has never had this issue.  Update the
documentation as well to reflect the new default.


Former-commit-id: 147407dd93e3f7c60f02209565357441bc577c94