docs: add guidelines for localization
We'd like to make it easier for folks to write code that is easy to
localize and to have localizations that are simple and intuitive.  Let's
add some documentation on how to write good localized strings, how to
perform the extraction of strings, and general advice both to
contributors of code and translations.


Former-commit-id: 4db7278965df1ca422c823d6b3f2ea0c02d64a53