t: add migrate tests for incorrect gitattrs perms
We currently generate incorrect duplicate .gitattributes
entries when performing an import or export migration if
an existing .gitattributes file has execute file permissions.

As we intend to fix this behaviour, we first add a pair
of tests for it, which check for a single .gitattributes
file after migration and therefore will fail until we
resolve the problem.


Former-commit-id: a7d666a7ef65bbcee63c4e8401c5483b5bb1382a