commands,git,t: fail migration on gitattrs symlink
If during a migration operation a .gitattributes blob is found
which needs to be modified by the migration, and it is a symbolic
link rather than a regular blob object, we halt the migration
with an error.

We add tests for the import, export, and info sub-commands which
exercise this new behaviour, including with and without the --fixup
option for the import and info sub-commands.

Note that this is an unexpected condition because .gitattributes
files should never be symbolic links, as Git itself will complain
about them with "Too many levels of symbolic links" warning messages.


Former-commit-id: 774ab58fa78d40e948740f8a08c52cb02a01a6ac