commands/command_ext.go: translate ext header only
In commit 1bfbcd60ac15287cb71a4d216c0afd9a76e3697d [formerly 1ddc2f2662b71d365ca7934c39b4ae56a60b5673] of
PR #4781 text translation support was added for the
"git lfs ext" command and a multi-line message was
concatenated.  However, the latter part of this text should
be fixed for all users, since it displays a listing of
the extension's settings values using defined terms.

We therefore separate the first line of the message and
let that be translated, as it contains the language-specific
header text.


Former-commit-id: a2d8e90d6eb37eb16a9f8532b02750f8bf9b88a2