git: remove git message prefixes
Several error messages output by the git package currently begin
with a "git:" prefix, which is unlike any other error messages
from that package, or anywhere else, for that matter.

We can therefore simplify our message strings slightly by simply
removing this prefix from them, and in one case, we reword the
error message to provide the full text of the failing Git command
(git for-each-ref), which is how other Git errors are reported.

Note that these messages are not yet passed as translation
strings, but we will address that issue in a subsequent commit.


Former-commit-id: 880eb15b2a887cf6341d1ebe3cb0db10afbb855e