git/git.go: use lowercase git command in message
We convert one message output by the git package to use
the lowercase "git" binary name instead of the capitalized
variant, which will make this message consistent with how
other Git commands are reported in error messages elsewhere.

Note that this message is not yet passed as a translation
string, but we will address that issue in a subsequent commit.


Former-commit-id: d02ac26e9db944a69f6f1d874490e9e6f2922b02