git/gitattr: remove git/gitattr prefix
One error message output by the git/gitattr package currently
begins with a "git/gitattr:" prefix, which is unlike other
error messages from that package, except for one where the
entire message consists of just the package name, and which
we will replace in a subsequent commit.

We can therefore simplify the message string with the prefix
simply removing the prefix.

Note that this message is not yet passed as a translation
string, but we will address that issue in a subsequent commit.


Former-commit-id: 7ed3787d2065442efb5ab4e34370e9e579e985db