lfs*,locking,tq: revise ancillary error messages
We rephrase some error message strings which are used when
wrapping other errors so they are clearer and more consistent
with other messages.  Note that these strings will be
prepended to the wrapper errors' messages.

In the lfs/diff_index_scanner.go file in particular we
rephrase the additional message to include a full Git
command ("git diff-index"), which is similar to how errors
are reported in the git package.


Former-commit-id: 0f2f1cdd9f884cc99cc523da7d6af9335d509786