commands/command_migrate_info.go: exit loop early
As we do in a similar scan through tree entries for a
.gitattributes entry in the "import" sub-command of the
"git lfs migrate" command, in the trackedFromAttrs() function,
we can also exit the loop after finding the entry in the
equivalent loop in the "info" sub-command.

We can also be more selective in our matching to ensure we
only look for blob entries.  And we should also match
case-insensitively on the .gitatttributes name.

h/t to bk2204 on PR review.


Former-commit-id: 439950b6b4c750df374a183dfbd4fc5d2820958a