commands,t: reverse ref args for fsck --objects
The "git lfs fsck --objects" command currently handles refs
expressed on the command line in two-dot range notation backwards;
this is due to the fact that we pass the first ref (identifying
everything to be excluded) as the "left" argument to the
gitscanner.ScanRefRange(), and the second ref as the "right"
argument.  That would seem to be correct, except that in the
gitscanner methods, "left" is meant to define all the commits to
include, while "right" is meant to define the ones to exclude.

This terminology likely stems from the way refs are passed to the
pre-push hook, with the local one coming first (on the left), and
the remote one coming second.  Since in that context we want to
exclude all the objects already on the remote, "right" (remote)
becomes the "exclude" ref, while "left" (local) becomes the
"include" one.

We can resolve this for the "git lfs fsck --objects" command
simply by reversing the two arguments.  Note that we don't need
to reverse the arguments for the --pointers command option,
even though it passes them in the same "backwards" order to the
gitscanner.ScanRefRangeByTree() method, because that method
happens to reverse its "left" and "right" arguments when it calls
scanRefsByTree().  We will address this counter-intuitive design
in a subsequent commit.

We also expand the "fsck operates on specified refs" test so that
it explicitly checks both the --pointers and --objects options,
checking the former against a commit with bad Git LFS pointer
objects, and checking the latter against a different commit with
an object file with incorrect contents.  This second new check
would fail without the reversal of the "left" and "right" arguments
to the gitscanner.ScanRefRange() method made in this commit.


Former-commit-id: f4c5d895998d26b936bfbfc9fd9ae204ade50475