smudge: honor GIT_LFS_SKIP_SMUDGE with checkout-index
In most cases, users use a relatively modern version of Git which
supports the filter-process code and delayed smudging.  This is valuable
because it makes things much faster. However, it also uses a different
code path from the non-delayed path. This non-delayed path is also used
by git checkout-index.

The non-delayed path doesn't work properly, however, if the data is
already on disk.  In such a case, we simply ask the smudge filter not to
download the data, but since it's already on disk, we feed it out
anyway.

Let's solve this by simply not invoking the filter and emitting the
pointer if we've asked to skip.  We do the same thing if the file
doesn't match the filter, since the comment at the top of the function
states that's what's supposed to happen here as well.

Once we've done that, the variable download is always true, so remove
all the code which considers that it might be false to simplify the code
somewhat.


Former-commit-id: 56b18085ab0bc731ef139d7a51c2b358349cc4b9