Call out destructive command in README
While it is the user's responsiblity to read the documentation and know
what impact a command will have that they choose to run on their
machine, I believe it would be helpful to draw more attention to the
destructive commands that are featured in the README that cause Git to
rewrite its commit objects.

Many users will only skim the README, and the migration command that
addresses a common use case should be called out more than it is for its
side effects, in my opinion.


Former-commit-id: 8aca97755bc6ad4e457460b0fd5cd6697cff4446