.github/workflows/*.yml: build on Windows 2019
The "gem install ronn" step in our Windows CI workflow currently
fails on newer Windows Server 2022 GitHub Actions runners, due to
possibly some missing C headers or libraries.

Ideally we would continue to use the "windows-latest" runners and
adjust the step which runs the now-deprecated "actions/setup-ruby"
action so that it runs "ruby/setup-ruby" instead.  However, doing
so introduces a set of problems whereby a different Bash shell is
used and our PATH environment variable settings are not respected,
the result being that when our tests run "git lfs" the system
default Git LFS binary is found by Git and executed instead of the
newly-built git-lfs.exe which we want to test, leading to many
failures.

So, for the time being, we resort to fixing our Windows CI and
release runners to use Windows Server 2019, until the problems can be
sorted out upstream.


Former-commit-id: f4ab4e7445c04fe5d663fc79ddb2ebf30a5a870f