script/windows-installer: drop old uninstall logic
In PR #875 in 2016 the Windows installer was changed from NSIS
to Inno Setup, and so logic was added in commit
aa36f2df2f0dd312d937d4e1bc925106eaf28e26 [formerly 1651e55ce5cc37c69bf5cb3bf92e626b83ad5e0a] to silently try to run
the old NSIS git-lfs-uninstaller.exe binary if it existed in
the same directory as git.exe, as found using the normal PATH
environment variables.

As all users of Git LFS should have upgraded to a recent version
installed with the Inno Setup installer, we can simply drop this
additional logic along with the GetExistingGitInstallation()
function of which it was the sole caller.

Co-authored-by: Johannes Schindelin <Johannes.Schindelin@gmx.de>

Former-commit-id: e06653d65e2afe88f6c8397f97155ce103234507